Added functionality for performing kmeans. Algorithm can put non clustered businesspoints into their appropriate initial clusters.

+++ Yelp_KMeans/CoffeeShops/src/app/DataPoint.java
@@ -3,9 +3,9 @@ package app;
 public class DataPoint{
 	 private int id;
 	 private String word;
-	 private int count;
+	 private double count;
 	 private String business;
-	 public DataPoint(int i, String w, int c, String b){
+	 public DataPoint(int i, String w, double c, String b){
 		 id = i;
 		 word = w;
 		 count = c;
@@ -17,7 +17,7 @@ public class DataPoint{
 	 public String getWord(){
 		 return word;
 	 }
-	 public int getCount(){
+	 public double getCount(){
 		 return count;
 	 }
 	 public String getBusiness(){
